// src/api/generated/model/mgmtRolloutResponseBody.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtRolloutResponseBodyTotalTargetsPerStatus } from './mgmtRolloutResponseBodyTotalTargetsPerStatus';
import type { MgmtRolloutResponseBodyType } from './mgmtRolloutResponseBodyType';
import type { Links } from './links';

/**
 * **_links**:
* **start** - Link to start the rollout in sync mode
* **pause** - Link to pause a running rollout
* **triggerNextGroup** - Link for triggering next rollout group on a running rollout
* **resume** - Link to resume a paused rollout
* **groups** - Link to retrieve the groups a rollout
* **approve** - Link to approve a rollout
* **deny** - Link to deny a rollout
* **distributionset** - The link to the distribution set

 */
export interface MgmtRolloutResponseBody {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  name: string;
  description?: string;
  /** Rollout id */
  id: number;
  /** Target filter query language expression */
  targetFilterQuery?: string;
  /** The ID of distribution set of this rollout */
  distributionSetId?: number;
  /** The status of this rollout */
  status: string;
  /** The total targets of a rollout */
  totalTargets: number;
  /** The total targets per status */
  totalTargetsPerStatus?: MgmtRolloutResponseBodyTotalTargetsPerStatus;
  /** The total number of groups created by this rollout */
  totalGroups?: number;
  /** Start at timestamp of Rollout */
  startAt?: number;
  /** Forcetime in milliseconds */
  forcetime?: number;
  /** Deleted flag, used for soft deleted entities */
  deleted?: boolean;
  /** The type of this rollout */
  type?: MgmtRolloutResponseBodyType;
  /** Weight of the resulting Actions */
  weight?: number;
  /** If this rollout is dynamic or static */
  dynamic?: boolean;
  approvalRemark?: string;
  approveDecidedBy?: string;
  _links?: Links;
}